feat: add feature to create, edit and delete list of seller products

+++ product-service/src/service/product-service.ts
@@ -6,6 +6,7 @@ import { AppValidationError } from "../util/errors";
 import { ProductInput } from "../dto/product-input";
 import { CategoryRepository } from "../repository/category-repository";
 import { ServiceInput } from "../dto/service-input";
+import { AuthUser } from "../util/auth";
 
 export class ProductService {
   _repository: ProductRepository;
@@ -17,12 +18,33 @@ export class ProductService {
     return ErrorResponse(404, new Error('method not allowed'));
   }
 
+  async authorizedUser(user_id: number, productId: string) {
+    const product = await this._repository.getProductById(productId);
+    if (!product) return false;
+    return Number(user_id) === Number(product.seller_id);
+  }
+
   async createProduct(event: APIGatewayEvent) {
-    const input = plainToClass(ProductInput, JSON.parse(event.body!));
+    // validate authorize user
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(
+        403,
+        "you need to join the seller program to create product"
+      );
+    }
+
+    const input = plainToClass(ProductInput, event.body);
     const error = await AppValidationError(input);
     if (error) return ErrorResponse(404, error);
 
-    const data = await this._repository.createProduct(input);
+    const data = await this._repository.createProduct({
+      ...input,
+      seller_id: user.user_id,
+    });
 
     await new CategoryRepository().addItem({
       id: input.category_id,
@@ -63,12 +85,28 @@ export class ProductService {
   }
 
   async editProduct(event: APIGatewayEvent) {
+    // validate authorize user
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(
+        403,
+        "you need to join the seller program to manage product"
+      );
+    }
+
     const productId = event.pathParameters?.id;
     if (!productId) return ErrorResponse(403, "please provide product id");
 
-    const input = plainToClass(ProductInput, JSON.parse(event.body!));
+    const input = plainToClass(ProductInput, event.body);
     const error = await AppValidationError(input);
     if (error) return ErrorResponse(404, error);
+
+    const isAuthorized = await this.authorizedUser(user.user_id, productId);
+    if (!isAuthorized)
+      return ErrorResponse(403, "you are not authorized to edit this product");
 
     input.id = productId;
     const data = await this._repository.updateProduct(input);
@@ -77,13 +115,30 @@ export class ProductService {
   }
 
   async deleteProduct(event: APIGatewayEvent) {
+    // validate authorize user
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(
+        403,
+        "you need to join the seller program to manage product"
+      );
+    }
+
     const productId = event.pathParameters?.id;
     if (!productId) return ErrorResponse(403, "please provide product id");
+
+    const isAuthorized = await this.authorizedUser(user.user_id, productId);
+    if (!isAuthorized)
+      return ErrorResponse(403, "you are not authorized to delete this product");
 
     const { category_id, deleteResult } = await this._repository.deleteProduct(
       productId
     );
-    await new CategoryRepository().addItem({
+
+    await new CategoryRepository().removeItem({
       id: category_id,
       products: [productId],
     });
@@ -91,7 +146,6 @@ export class ProductService {
   }
 
   // http calls // later stage will convert this thing to RPC & Queue
-
   async handleQueueOperation(event: APIGatewayProxyEvent) {
     const input = plainToClass(ServiceInput, event.body);
     const error = await AppValidationError(input);
